agents/official/beeai-framework/chat/pyproject.toml
[project]
name = "chat"
version = "0.1.0"
description = "AI-powered conversational system with memory and tools."
authors = [
    { name = "IBM Corp." },
]
requires-python = ">=3.13,<4"
dependencies = [
    "acp-sdk==1.0.0",
    "beeai-framework[duckduckgo,wikipedia]~=0.1.31",
    "openinference-instrumentation-beeai>=0.1.6",
    "pydantic-settings>=2.9.0",
]

[project.scripts]
server = "chat.agent:run"

[build-system]
requires = [
    "hatchling",
]
build-backend = "hatchling.build"
